separates types from reducers
