+++ app/components/TimeEntry/TimeEntry.js
@@ -8,7 +8,7 @@ import style from './style.scss';
 import { startTracking, stopTracking } from '../../actions/toggl';
 import { secondsToHours } from '../../utils/helpers';
 import Timer from '../Timer/Timer';
-import type { TimeEntry as TimeEntryType } from '../../reducers/toggl';
+import type { TimeEntry as TimeEntryType } from '../../types';
 
 class TimeEntry extends Component {
   props: {
@@ -22,7 +22,7 @@ class TimeEntry extends Component {
     tag: ?[string],
     billable: boolean,
     color: ?string,
-    startTracking: (Object) => null,
+    startTracking: (TimeEntryType) => null,
     stopTracking: (number) => null,
     timeEntry: TimeEntryType
   };
